+++ authress_sdk/api/resource_permissions_api.py
@@ -8,14 +8,10 @@ import re
 # python 2 and python 3 compatibility library
 import six
 
-from authress_sdk.api_client import ApiClient
-
 
 class ResourcePermissionsApi(object):
-    def __init__(self, api_client=None):
-        if api_client is None:
-            api_client = ApiClient()
-        self.api_client = api_client
+    def __init__(self, api_client):
+      self.api_client = api_client
 
     def get_resources(self, **kwargs):
         """List resource configurations