+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This is the changelog for [Authress SDK](readme.md).
 ## 2.0 ##
 * Add support for users and groups at the statement level of access records.
 * Fix naming of some incorrectly named plural classes in the Access Record Models.
+* Rename `ApiClient` to `AuthressClient` and split out non-essential functionality to a separate HttpClient.
 
 ## 1.1 ##
 * Add the `TokenVerifier`